+++ elasticgraph-graphql/sig/elastic_graph/graphql/config.rbs
@@ -1,7 +1,7 @@
 module ElasticGraph
   class GraphQL
     class ConfigSupertype
-      extend ::ElasticGraph::Config::ClassMethods[Config]
+      extend Support::Config::ClassMethods[Config]
 
       attr_reader default_page_size: ::Integer
       attr_reader max_page_size: ::Integer
