+++ elasticgraph-query_registry/sig/elastic_graph/query_registry/config.rbs
@@ -1,7 +1,7 @@
 module ElasticGraph
   module QueryRegistry
     class ConfigSupertype
-      extend ::ElasticGraph::Config::ClassMethods[Config]
+      extend Support::Config::ClassMethods[Config]
 
       attr_reader path_to_registry: ::String
       attr_reader allow_unregistered_clients: bool
